+++ R/zotnegbin.R
@@ -245,7 +245,7 @@ zotnegbin <- function(nSim = 1000, epsSim = 1e-8,
         alpha  <-  alphaLink(eta[, 2], inverse = TRUE)
 
         -sum(weight * (lgamma(y + 1 / alpha) - lgamma(1 / alpha) -
-        log(factorial(y)) - (y + 1 / alpha) * log(1 + lambda * alpha) +
+        lgamma(y + 1) - (y + 1 / alpha) * log(1 + lambda * alpha) +
         y * log(lambda * alpha) - log(1 - (1 + lambda * alpha) ^ (-1 / alpha) - 
         lambda * (1 + lambda * alpha) ^ (-1 - 1 / alpha)))
         )
@@ -381,7 +381,7 @@ zotnegbin <- function(nSim = 1000, epsSim = 1e-8,
     
     # 
     # logLikFit <- (lgamma(y + 1/alpha) - lgamma(1/alpha) -
-    # log(factorial(y)) - (y + 1/alpha) * log(1+alpha * lambda) +
+    # lgamma(y + 1) - (y + 1/alpha) * log(1+alpha * lambda) +
     # y * log(lambda * alpha) - log(1 - (1+alpha * lambda) ^ (-1/alpha) -
     # lambda * ((1+alpha * lambda) ^ (-1-1/alpha))))
     #