+++ R/ztoinegbin.R
@@ -391,7 +391,7 @@ ztoinegbin <- function(nSim = 1000, epsSim = 1e-8, eimStep = 6,
               lambda * (1 + alpha * lambda) ^ (-1 / alpha - 1) / 
               (1 - (1 + lambda * alpha) ^ (-1 / alpha))) + 
               (1 - z) * (log(1 - omega) + lgamma(y + 1 / alpha) - 
-              lgamma(1 / alpha) - log(factorial(y)) - 
+              lgamma(1 / alpha) - lgamma(y + 1) - 
               (y + 1 / alpha) * log(1 + lambda * alpha) + y * log(lambda * alpha) - 
               log(1 - (1 + lambda * alpha) ^ (-1 / alpha)))))
             },