+++ R/ztpoisson.R
@@ -77,7 +77,7 @@ ztpoisson <- function(lambdaLink = c("log", "neglog"),
       function(beta) {
         lambda <- lambdaLink((as.matrix(X) %*% beta)[, 1], inverse = TRUE)
         
-        -sum(weight * (y * log(lambda) - log(exp(lambda) - 1) - log(factorial(y))))
+        -sum(weight * (y * log(lambda) - log(exp(lambda) - 1) - lgamma(y + 1)))
       },
       function(beta) {
         eta <- as.matrix(X) %*% beta
@@ -116,10 +116,40 @@ ztpoisson <- function(lambdaLink = c("log", "neglog"),
 
   devResids <- function(y, eta, wt, ...) {
     lambda <- lambdaLink(eta[, 1], inverse = TRUE)
-    idealLambda <- ifelse(y > 1, lamW::lambertW0(-y * exp(-y)) + y, 0)
+    
+    idealLambda <- tryCatch(
+      expr = {
+        suppressWarnings(
+          ifelse(y > 1, lamW::lambertW0(-y * exp(-y)) + y, 0)
+        )
+      },
+      error = function (e) {
+        warning("Deviance residuals could not have been computed and zero vector will be returned instead.", call. = FALSE)
+        NULL
+      }
+    )
+    if (is.null(idealLambda)) {
+      return(rep(0, length(y)))
+    }
+    
     
     logL <- y * log(lambda) - lambda - log(1 - exp(-lambda))
     logSat <- y * ifelse(y > 1, log(idealLambda), 0) - idealLambda - ifelse(y > 1, log(1 - exp(-idealLambda)), 0)
+    
+    diff <- (y * log(lambda) - lambda - log(1 - exp(-lambda)) - 
+    y * ifelse(y > 1, log(idealLambda), 0) - 
+    idealLambda - ifelse(y > 1, log(1 - exp(-idealLambda)), 0)
+    )
+    
+    if (any(diff > 0)) {
+      warning(paste0(
+        "Some of differences between log likelihood in sautrated model",
+        " and fitted model were positive which idicates either:\n",
+        "(1): A very good model fitt or\n",
+        "(2): Incorrect computation of saturated model",
+        "\nDouble check deviance before proceeding"
+      ))
+    }
     
     ### Here we take pairwise minimum because in specific situations
     ### lambda and idealLambda are so close for some units that